ifound_api/requirements.txt
aiofiles==0.7.0
alembic==1.6.5
asgiref==3.4.1
autopep8==1.7.0
backcall==0.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
config==0.5.1
decorator==5.1.0
ecdsa==0.17.0
elastic-transport==8.1.2
elasticsearch==8.3.1
fastapi==0.68.0
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
importlib-metadata==4.6.4
invoke==1.7.1
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.0
loguru==0.5.3
Mako==1.1.4
mariadb==1.0.7
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.26
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.21
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.22
starkbank==2.13.0
starkbank-ecdsa==1.0.0
starlette==0.14.2
toml==0.10.2
traitlets==5.1.1
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
XlsxWriter==3.0.1
zipp==3.5.0
